refactor: extract single resizable constants
